# Disable using various workers in tests for FastText