Current version of FastText is pure-python,
so CPU-bound processing with multiple workers in it is a bad idea.
Using workers=12 in a some tests increased execution time of tests.