@menshikh-iv 2017-11-02
Thanks @horpto, can you compare the time of `test_fasttext` before/after?
CC: @chinmayapancholi13 WDYT about this change?