@menshikh-iv 2017-11-02
I run each test 3 times and averaged the results.

| Test | Before (sec) | After (sec) | Speedup (x)|
|------|--------------|-------------|------------|
| test_cbow_hs_against_wrapper | 115.7 | 52.5 | 2.2 |
| test_sg_hs_against_wrapper | 189.6 | 92.1 | 2.0 |
| test_sg_hs_online | 23.5 | 13.1 | 1.7 |
| test_sg_neg_online | 29.9 | 16.8 | 1.8 |
| test_cbow_hs_online | 18.0 | 9.0 | 2.0 |
| test_cbow_neg_online | 20.3 | 11.1| 1.8 |

Nice speedup for slowest tests, good job @horpto!